Adicionando relacionamentos e corrigindo alguns detalhes

+++ dot_net_api/Models/Produto.cs
@@ -1,10 +1,11 @@
+using System;
 namespace dot_net_api.Models
 {
     public class Produto
     {
         public int Id { get; set; }
         public string Nome { get; set; }
-        public decimal Preco { get; set; }
+        public Double Preco { get; set; }
         public Categoria Categoria { get; set; }
         public int CategoriaId { get; set; }
     }
